+++ gaitmap/trajectory_reconstruction/trajectory_methods/rts_kalman.py
@@ -1,5 +1,7 @@
 """An error state kalman filter with Rauch-Tung-Striebel smoothing fo estimating trajectories."""
-from typing import Union, TypeVar, Optional
+import warnings
+from typing import TypeVar, Optional
+from typing import Union
 
 import numpy as np
 import pandas as pd
@@ -7,10 +9,16 @@ from numba import njit
 from scipy.spatial.transform import Rotation
 
 from gaitmap.base import BaseTrajectoryMethod
+from gaitmap.utils.array_handling import bool_array_to_start_end_array
 from gaitmap.utils.consts import GRAV_VEC
 from gaitmap.utils.consts import SF_GYR, SF_ACC, GF_POS, GF_VEL
 from gaitmap.utils.datatype_helper import is_single_sensor_data, SingleSensorData
-from gaitmap.utils.fast_quaternion_math import quat_from_rotvec, multiply, rotate_vector
+from gaitmap.utils.fast_quaternion_math import (
+    quat_from_rotvec,
+    multiply,
+    rotate_vector,
+    normalize,
+)
 from gaitmap.utils.static_moment_detection import find_static_samples
 
 Self = TypeVar("Self", bound="RtsKalman")
@@ -62,6 +70,20 @@ class RtsKalman(BaseTrajectoryMethod):
     level_walking_variance
         The variance of the noise of the measured position during a level walking update.
         Should typically be very small.
+    zupt_orientation_update
+        Flag to control, if the orientation of the IMU should be actively corrected using acceleration information
+        during ZUPT.
+        If True, the direction of gravity is guessed based on the direction of the acc-vector.
+        This direction is then aligned with the expected gravity direction.
+        This is disabled by default, as it can lead to unexpected results, if there is a lot of movement during the
+        mid stance.
+        However, depending on the dataset, this can be very benifical.
+
+        .. warning:: The support for orientation updates is still experimental and not fully validated on real data.
+
+    zupt_orientation_error_variance
+        The variance of the noise of the measured orientation during the ZUPT.
+        This might be comparably high, because we still expcet some variation in the accelerometer.
     zupt_window_length_s
         Length of the window used in the default method to find ZUPTs.
         If the value is too small at least a window of 2 samples is used.
@@ -84,6 +106,8 @@ class RtsKalman(BaseTrajectoryMethod):
     covariance_
         The covariance matrices of the kalman filter after smoothing.
         They can be used as a measure of how good the filter worked and how accurate the results are.
+    zupts_
+        2D array indicating the start and the end samples of the detected ZUPTs for debug porpuses.
 
     Other Parameters
     ----------------
@@ -153,9 +177,14 @@ class RtsKalman(BaseTrajectoryMethod):
     level_walking_variance: float
     zupt_window_length_s: float
     zupt_window_overlap_s: Optional[float]
+    zupt_orientation_update: bool
+    zupt_orientation_error_variance: float
+
     data: SingleSensorData
     sampling_rate_hz: float
+
     covariance_: pd.DataFrame
+    zupts_: np.ndarray
 
     def __init__(
         self,
@@ -166,6 +195,8 @@ class RtsKalman(BaseTrajectoryMethod):
         orientation_error_variance: float = 10e-2,
         level_walking: bool = True,
         level_walking_variance: float = 10e-8,
+        zupt_orientation_update: bool = False,
+        zupt_orientation_error_variance: float = 10e-1,
         zupt_window_length_s: float = 0.05,
         zupt_window_overlap_s: Optional[float] = None,
     ):
@@ -176,6 +207,8 @@ class RtsKalman(BaseTrajectoryMethod):
         self.orientation_error_variance = orientation_error_variance
         self.level_walking = level_walking
         self.level_walking_variance = level_walking_variance
+        self.zupt_orientation_update = zupt_orientation_update
+        self.zupt_orientation_error_variance = zupt_orientation_error_variance
         self.zupt_window_length_s = zupt_window_length_s
         self.zupt_window_overlap_s = zupt_window_overlap_s
 
@@ -200,6 +233,12 @@ class RtsKalman(BaseTrajectoryMethod):
         self.sampling_rate_hz = sampling_rate_hz
         initial_orientation = self.initial_orientation
 
+        if self.zupt_orientation_update is True:
+            warnings.warn(
+                "Support for zupt orientation updates (`zupt_orientation_update=True`) is still "
+                "experimental and not properly validated on large datasets."
+            )
+
         is_single_sensor_data(self.data, frame="sensor", raise_exception=True)
         if isinstance(initial_orientation, Rotation):
             initial_orientation = Rotation.as_quat(initial_orientation)
@@ -211,13 +250,17 @@ class RtsKalman(BaseTrajectoryMethod):
         covariance = np.copy(process_noise)
 
         # measure noise
-        meas_noise = self.zupt_variance * np.eye(4)
-        if not self.level_walking:
-            meas_noise[3, 3] = 0.0
+        # TODO: Set different zupt noises for ori and vel?
+        meas_noise = np.zeros((6, 6))
+        meas_noise[0:3, 0:3] = np.eye(3) * self.zupt_variance
+        if self.level_walking is True:
+            meas_noise[3, 3] = self.level_walking_variance
+        if self.zupt_orientation_update:
+            meas_noise[4:6, 4:6] = np.eye(2) * self.zupt_orientation_error_variance
 
         gyro_data = np.deg2rad(data[SF_GYR].to_numpy())
         acc_data = data[SF_ACC].to_numpy()
-        zupts = self.find_zupts(gyro_data)
+        zupts = self.find_zupts(gyro_data, self.sampling_rate_hz)
 
         states, covariances = _rts_kalman_update_series(
             acc_data,
@@ -229,6 +272,7 @@ class RtsKalman(BaseTrajectoryMethod):
             process_noise,
             zupts,
             self.level_walking,
+            self.zupt_orientation_update,
         )
         self.position_ = pd.DataFrame(states[0], columns=GF_POS)
         self.position_.index.name = "sample"
@@ -247,23 +291,39 @@ class RtsKalman(BaseTrajectoryMethod):
             "y_ori_cov",
             "z_ori_cov",
         ]
-        self.covariance_ = pd.concat(
-            [pd.DataFrame(cov, columns=covariance_cols, index=covariance_cols) for cov in covariances],
-            keys=range(len(covariances)),
-        )
+        covariance_cols = pd.MultiIndex.from_product((covariance_cols, covariance_cols))
+        covariances = covariances.reshape((covariances.shape[0], -1))
+        self.covariance_ = pd.DataFrame(covariances, columns=covariance_cols)
+
+        self.zupts_ = bool_array_to_start_end_array(zupts)
         return self
 
-    def find_zupts(self, gyro):
-        """Find the ZUPT samples based on the gyro measurements."""
-        window_length = max(2, round(self.sampling_rate_hz * self.zupt_window_length_s))
+    def find_zupts(self, gyro, sampling_rate_hz: float):
+        """Find the ZUPT samples based on the gyro measurements.
+
+        Parameters
+        ----------
+        gyro
+            gyro in rad/s
+        sampling_rate_hz
+            sampling rate of the gyro data
+
+        Returns
+        -------
+        zupt_array
+            array of length gyro with True and False indicating a ZUPT.
+
+        """
+        window_length = max(2, round(sampling_rate_hz * self.zupt_window_length_s))
         zupt_window_overlap_s = self.zupt_window_overlap_s
         if zupt_window_overlap_s is None:
             window_overlap = int(window_length // 2)
         else:
-            window_overlap = round(self.sampling_rate_hz * zupt_window_overlap_s)
-        return find_static_samples(
+            window_overlap = round(sampling_rate_hz * zupt_window_overlap_s)
+        zupts = find_static_samples(
             gyro, window_length, self.zupt_threshold_dps * (np.pi / 180), "maximum", window_overlap
         )
+        return zupts
 
 
 @njit()
@@ -285,8 +345,17 @@ def _rts_kalman_motion_update(acc, gyro, orientation, position, velocity, sampli
 
 
 @njit()
-def _rts_kalman_forward_pass(
-    accel, gyro, initial_orientation, sampling_rate_hz, meas_noise, covariance, process_noise, zupts, level_walking
+def _rts_kalman_forward_pass(  # noqa: too-many-statements, too-many-branches
+    accel,
+    gyro,
+    initial_orientation,
+    sampling_rate_hz,
+    meas_noise,
+    covariance,
+    process_noise,
+    zupts,
+    level_walking,
+    orientation_correction,
 ):
     prior_covariances = np.empty((accel.shape[0] + 1, 9, 9))
     posterior_covariances = np.empty((accel.shape[0] + 1, 9, 9))
@@ -308,10 +377,38 @@ def _rts_kalman_forward_pass(
     transition_matrix = np.eye(9)
     transition_matrix[:3, 3:6] = np.eye(3) / sampling_rate_hz
 
-    meas_matrix = np.zeros((4, 9))
-    meas_matrix[:3, 3:6] = np.eye(3)
-    if level_walking:
-        meas_matrix[3, 2] = 1.0
+    gravity = normalize(GRAV_VEC)
+
+    # During zupt we measure 6 values:
+    # 1-3: v_{x,y,z} = 0
+    # 4 : p_z = 0
+    # 5 : angle(global_acc, global_grav) around x-axis = 0
+    # 6 : angle(global_acc, global_grav) around y-axis = 0
+    #
+    # The values 1-4 are directly part of the stater space.
+    # This means we have a trivial measurement function h and jacobian H
+    # Value 5 is the angle between the local z-axis and the global z-axis.
+    # This can be represented as the second euler angle in a 3-1-3 euler angel configuration.
+    # Hence we can calculate it from the quaternion components.
+    # See notes on the full calculation in the ZUPT update code.
+    zupt_measurement = np.zeros(6)
+    # meas_jacob dh/d(\delta x) maps from measurement space into the error space
+    # Because we directly observe the values from our error vector without any conversion, it just consists of 1 in the
+    # right places.
+    meas_jacob = np.zeros((6, 9))
+    # The meas_func is used to mask only the correction values we want to have.
+    meas_func = np.zeros(6)
+    # Zero Velocity update
+    meas_func[0:3] = 1
+    meas_jacob[0:3, 3:6] = np.eye(3)
+    if level_walking is True:
+        # Zero elevation update
+        meas_jacob[3, 2] = 1
+        meas_func[3] = 1
+    if orientation_correction is True:
+        # Angle error update
+        meas_func[4:6] = 1
+        meas_jacob[4:6, 6:8] = np.eye(2)
 
     for i, zupt in enumerate(zupts):
         acc = np.ascontiguousarray(accel[i])
@@ -336,18 +433,38 @@ def _rts_kalman_forward_pass(
         )
 
         # correct the error state if the current sample is marked as a zupt, this is the update step
+        # Note, that the error state is not used to correct the nominal state at this point.
+        # We are just correcting the error state.
+        # Fusing with the nominal state is performed in the correction step of the smoothing.
         if zupt:
-            innovation = np.append(velocity, position[2]) - meas_matrix @ prior_error_states[i + 1]
-            innovation_cov = meas_matrix @ prior_covariances[i + 1] @ meas_matrix.T + meas_noise
-            inv = np.zeros((4, 4))
-            if level_walking:
-                inv[:, :] = np.linalg.inv(innovation_cov)
-            else:
-                inv[:3, :3] = np.linalg.inv(innovation_cov[:3, :3])
-            gain = prior_covariances[i + 1] @ meas_matrix.T @ inv
-            posterior_error_states[i + 1, :] = prior_error_states[i + 1] + gain @ innovation
+            meas_space_error = meas_jacob @ prior_error_states[i + 1]
 
-            factor = np.eye(covariance.shape[0]) - gain @ meas_matrix
+            # velocity error
+            zupt_measurement[0:3] = velocity
+            if level_walking:
+                # z-position error
+                zupt_measurement[3] = position[2]
+            # orientation error
+            if orientation_correction:
+                # Find the angles between rotated acc and gravity
+                norm_acc = normalize(rotated_acc)
+                cross = np.cross(norm_acc, gravity)
+                dot = np.dot(norm_acc, gravity)
+                zupt_measurement[4] = np.arctan2(np.dot(cross, np.array([1.0, 0, 0])), dot)
+                zupt_measurement[5] = np.arctan2(np.dot(cross, np.array([0, 1.0, 0])), dot)
+
+            innovation = meas_func.copy()
+            # Instead of using the calculated error, we calculate how much the error has increased since the last time
+            # step.
+            # This value is than used to calculate the change in covariance.
+            # This is needed, because we do not apply the innovation to the actual state as we would do in a regular
+            # kalman filter.
+            # The error is just summed up and only applied to the state in the final smoothing pass.
+            innovation *= zupt_measurement - meas_space_error
+            innovation_cov = meas_jacob @ prior_covariances[i + 1] @ meas_jacob.T + meas_noise
+            gain = prior_covariances[i + 1] @ meas_jacob.T @ np.linalg.pinv(innovation_cov)
+            posterior_error_states[i + 1, :] = prior_error_states[i + 1] + gain @ innovation
+            factor = np.eye(covariance.shape[0]) - gain @ meas_jacob
             covariance_adj = factor @ prior_covariances[i + 1] @ factor.T
             posterior_covariances[i + 1, :, :] = covariance_adj + gain @ meas_noise @ gain.T
         else:
@@ -395,10 +512,28 @@ def _rts_kalman_correction_pass(positions, velocities, orientations, corrected_e
 
 @njit(cache=True)
 def _rts_kalman_update_series(
-    acc, gyro, initial_orientation, sampling_rate_hz, meas_noise, covariance, process_noise, zupts, level_walking
+    acc,
+    gyro,
+    initial_orientation,
+    sampling_rate_hz,
+    meas_noise,
+    covariance,
+    process_noise,
+    zupts,
+    level_walking,
+    orientation_correction,
 ):
     forward_eskf_results, forward_nominal_states = _rts_kalman_forward_pass(
-        acc, gyro, initial_orientation, sampling_rate_hz, meas_noise, covariance, process_noise, zupts, level_walking
+        acc,
+        gyro,
+        initial_orientation,
+        sampling_rate_hz,
+        meas_noise,
+        covariance,
+        process_noise,
+        zupts,
+        level_walking,
+        orientation_correction,
     )
     corrected_error_states, corrected_covariances = _rts_kalman_backward_pass(*forward_eskf_results)