+++ tests/test_trajectory_reconstruction/test_trajectory_methods/test_rts_kalman.py
@@ -1,12 +1,13 @@
 import numpy as np
 import pandas as pd
 import pytest
-from numpy.testing import assert_array_almost_equal
 from numpy.linalg import norm
+from numpy.testing import assert_array_almost_equal
 
-from gaitmap.base import BaseType, BaseOrientationMethod
+from gaitmap.base import BaseType
 from gaitmap.trajectory_reconstruction import RtsKalman
 from gaitmap.utils.consts import SF_COLS
+from gaitmap.utils.rotations import find_unsigned_3d_angle
 from tests.mixins.test_algorithm_mixin import TestAlgorithmMixin
 from tests.test_trajectory_reconstruction.test_trajectory_methods.test_trajectory_method_mixin import (
     TestTrajectoryMethodMixin,
@@ -25,12 +26,33 @@ class TestMetaFunctionality(TestAlgorithmMixin):
         )
         return kalman_filter
 
+    def test_experimental_warning(self):
+        fs = 10
+        sensor_data = np.repeat(np.array([0.0, 0.0, 9.81, 0.0, 0.0, 0.0])[None, :], fs, axis=0)
+        sensor_data = pd.DataFrame(sensor_data, columns=SF_COLS)
+
+        rts = RtsKalman(zupt_orientation_update=True)
+
+        with pytest.warns(UserWarning) as w:
+            rts.estimate(sensor_data, fs)
+
+        assert "experimental" in str(w[0])
+
 
 class TestTrajectoryMethod(TestTrajectoryMethodMixin):
     __test__ = True
 
-    def init_algo_class(self) -> BaseOrientationMethod:
-        return RtsKalman()
+    default_kwargs = {}
+
+    @pytest.fixture(autouse=True, params=(True, False), ids=("ori_correction", "no_ori_correction"))
+    def ori_correction(self, request):
+        if "noori" in request.keywords and request.param is False:
+            pytest.skip()
+        self.default_kwargs = {"zupt_orientation_update": request.param}
+
+    def init_algo_class(self, **kwargs) -> RtsKalman:
+        kwargs = {**self.default_kwargs, **kwargs}
+        return RtsKalman(**kwargs)
 
     def test_covariance_output_format(self):
         test = self.init_algo_class()
@@ -39,9 +61,19 @@ class TestTrajectoryMethod(TestTrajectoryMethodMixin):
         sensor_data = pd.DataFrame(sensor_data, columns=SF_COLS)
         test.estimate(sensor_data, fs)
 
-        assert len(test.covariance_) == (len(sensor_data) + 1) * 9
-        assert len(test.covariance_.columns) == 9
-        assert test.covariance_.index.levshape == (len(sensor_data) + 1, 9)
+        assert test.covariance_.shape == (len(sensor_data) + 1, 9 * 9)
+
+    def test_zupt_output(self):
+        test = self.init_algo_class(zupt_threshold_dps=10, zupt_window_length_s=0.01)
+        fs = 15
+        sensor_data = np.repeat(np.array([0.0, 0.0, 9.81, 0.0, 0.0, 100.0])[None, :], 100, axis=0)
+        expected_zupts = [[0, 10], [30, 55], [85, 90]]
+        for z in expected_zupts:
+            sensor_data[slice(*z), -1] = 0
+        sensor_data = pd.DataFrame(sensor_data, columns=SF_COLS)
+        test.estimate(sensor_data, fs)
+
+        assert_array_almost_equal(expected_zupts, test.zupts_)
 
     def test_corrects_velocity_drift(self):
         """Check that ZUPTs correct a velocity drift and set velocity to zero."""
@@ -65,3 +97,23 @@ class TestTrajectoryMethod(TestTrajectoryMethodMixin):
         test.estimate(sensor_data, 10)
         assert test.position_.to_numpy()[4][2] < -0.8
         assert_array_almost_equal(test.position_.to_numpy()[-1], [0.0, 0.0, 0.0], decimal=10)
+
+    @pytest.mark.noori
+    def test_corrects_ori(self):
+        """Check that small errors in the orientation are correctly fixed."""
+        test = self.init_algo_class(zupt_orientation_update=True, level_walking=False)
+        accel_data = np.repeat(np.concatenate(([0.0, 0.0, 9.81], [40.0, 40.0, 0.0]))[None, :], 5, axis=0)
+        zupt_data = np.repeat(np.concatenate(([0.0, 0.0, 9.81], [0.0, 0.0, 0.0]))[None, :], 20, axis=0)
+        sensor_data = np.vstack((accel_data, zupt_data))
+        sensor_data = pd.DataFrame(sensor_data, columns=SF_COLS)
+        test.estimate(sensor_data, 10)
+        rotated_acc = test.orientation_object_[:-1].apply(sensor_data.filter(like="acc"))
+        # We calculate the angle to make check that this gets steadly reduced as soon as the ZUPT starts
+        # Note that the error actual goes through 0 and then increases again.
+        # This is likely because, we are correcting large angle values and we have a certain linearisation error.
+        # The angle will then converge to a small value
+        angle = find_unsigned_3d_angle(rotated_acc, np.array([0, 0, 1]))
+        diff_angle = np.diff(angle)
+        assert all(diff_angle[1:5] > 0)
+        assert all(diff_angle[5:10] < 0)
+        assert all(angle[6:] < angle[5])